+++ ui/src/layout/projectSiderbar.tsx
@@ -9,7 +9,7 @@ import PutAwayMemu from 'components/PutAwayMenu';
 import TextLimit from 'components/TextLimit';
 import Loading from 'components/Loading';
 import { IRouterParams, IProject, IEnvironment } from 'interfaces/project';
-import { TOGGLE_PATH, TARGETING_PATH, SEGMENT_PATH, SEGMENT_ADD_PATH, SEGMENT_EDIT_PATH, GET_STARTED_PATH, SETTING_PATH } from 'router/routes';
+import { TOGGLE_PATH, TARGETING_PATH, SEGMENT_PATH, SEGMENT_ADD_PATH, SEGMENT_EDIT_PATH, CONNECT_SDK_PATH, SETTING_PATH } from 'router/routes';
 import { SidebarContainer } from './hooks';
 import styles from './sidebar.module.scss';
 
@@ -64,7 +64,7 @@ const ProjectSiderbar = (props: IProps) => {
 
   useEffect(() => {
     const { path } = match;
-    if (path === TOGGLE_PATH || path === TARGETING_PATH || path === GET_STARTED_PATH) {
+    if (path === TOGGLE_PATH || path === TARGETING_PATH || path === CONNECT_SDK_PATH) {
       setSelectedItem('toggle');
     } else if (path === SEGMENT_PATH || path === SEGMENT_ADD_PATH || path === SEGMENT_EDIT_PATH) {
       setSelectedItem('segments');
