+++ ui/src/pages/connectSDK/components/constants.ts
