+++ ui/src/pages/connectSDK/index.module.scss
@@ -1,4 +1,5 @@
-.page {
+.connect-sdk {
+  padding: 0;
   margin: 0 20px;
 }
 
@@ -378,3 +379,4 @@
   margin-left: 8px;
   color: #556ee6;
 }
+
