+++ ui/src/pages/connectSDK/index.tsx
@@ -1,18 +1,19 @@
 import { useState, useEffect, useCallback } from 'react';
 import { FormattedMessage } from 'react-intl';
 import { useParams } from 'react-router-dom';
+import ProjectLayout from 'layout/projectLayout';
 import Loading from 'components/Loading';
-import StepFirst from '../StepFirst';
-import StepSecond, { SdkLanguage } from '../StepSecond';
-import StepThird from '../StepThird';
+import StepFirst from './components/StepFirst';
+import StepSecond, { SdkLanguage } from './components/StepSecond';
+import StepThird from './components/StepThird';
+import { ToggleReturnType } from './components/constants';
 import { saveDictionary, getFromDictionary } from 'services/dictionary';
 import { getSdkVersion } from 'services/misc';
 import { getToggleAccess, getToggleInfo, getToggleAttributes } from 'services/toggle';
-import { getProjectInfo } from 'services/project';
-import { getEnvironment } from 'services/project';
+import { getProjectInfo, getEnvironment } from 'services/project';
 import { IDictionary, IToggleInfo } from 'interfaces/targeting';
 import { IProject, IEnvironment, IRouterParams } from 'interfaces/project';
-import { ToggleReturnType } from '../constants';
+
 import styles from './index.module.scss';
 
 interface IStepDetail {
@@ -50,7 +51,7 @@ const JAVA_SDK_VERSION = 'java_sdk_version';
 const RUST_SDK_VERSION = 'rust_sdk_version';
 const ANDROID_SDK_VERSION = 'android_sdk_version';
 
-const Steps = () => {
+const ConnectSDK = () => {
   const [ currentStep, saveCurrentStep ] = useState<number>(2);
   const [ currentSDK, saveCurrentSDK ] = useState<SdkLanguage>('Java');
   const [ serverSdkKey, saveServerSDKKey ] = useState<string>('');
@@ -74,10 +75,11 @@ const Steps = () => {
     Promise.all([
       getFromDictionary<IDictionary>(key), 
       getToggleAttributes<string[]>(projectKey, environmentKey, toggleKey)
-     ]).then(res => {
+    ]).then(res => {
       saveIsStepLoading(false);
       if (res[0].success && res[0].data) {
         const savedData = JSON.parse(res[0].data.value);
+        console.log(savedData);
         if (savedData.step2.done) {
           saveCurrentStep(3);
           saveCurrentSDK(savedData.step1.sdk);
@@ -188,91 +190,93 @@ const Steps = () => {
   }, []);
 
   return (
-    <div className={styles.page}>
-      <div className={styles.intro}>
-        {
-          isInfoLoading ? <Loading /> : (
-            <>
-              <div className={styles['intro-header']}>
-                <span className={styles['intro-title']}>
-                  <FormattedMessage id='common.get.started.text' />
-                </span>
-              </div>
-              <div className={styles['intro-desc']}>
-                <FormattedMessage id='connect.description' />
-              </div>
-              <div className={styles['intro-info']}>
-                <div className={styles['card-item']}>
-                  <div className={styles['card-title']}>
-                    <FormattedMessage id='common.project.text' /> :
+    <ProjectLayout>
+      <div className={styles['connect-sdk']}>
+        <div className={styles.intro}>
+          {
+            isInfoLoading ? <Loading /> : (
+              <>
+                <div className={styles['intro-header']}>
+                  <span className={styles['intro-title']}>
+                    <FormattedMessage id='common.get.started.text' />
+                  </span>
+                </div>
+                <div className={styles['intro-desc']}>
+                  <FormattedMessage id='connect.description' />
+                </div>
+                <div className={styles['intro-info']}>
+                  <div className={styles['card-item']}>
+                    <div className={styles['card-title']}>
+                      <FormattedMessage id='common.project.text' /> :
+                    </div>
+                    <div className={styles['card-value']}>
+                      { projectName }
+                    </div>
                   </div>
-                  <div className={styles['card-value']}>
-                    { projectName }
+                  <div className={styles['card-item']}>
+                    <div className={styles['card-title']}>
+                      <FormattedMessage id='common.environment.text' /> :
+                    </div>
+                    <div className={styles['card-value']}>
+                      { environmentName }
+                    </div>
+                  </div>
+                  <div className={styles['card-item']}>
+                    <div className={styles['card-title']}>
+                      <FormattedMessage id='common.toggle.text' /> :
+                    </div>
+                    <div className={styles['card-value']}>
+                      <FormattedMessage id='connect.first.toggle.view.left' />
+                      <span className={styles['toggle-name']}>{ toggleName }</span>
+                      <FormattedMessage id='connect.first.toggle.view.right' />
+                      <span className={styles['toggle-key']}>{ toggleKey }</span>
+                    </div>
                   </div>
                 </div>
-                <div className={styles['card-item']}>
-                  <div className={styles['card-title']}>
-                    <FormattedMessage id='common.environment.text' /> :
-                  </div>
-                  <div className={styles['card-value']}>
-                    { environmentName }
-                  </div>
-                </div>
-                <div className={styles['card-item']}>
-                  <div className={styles['card-title']}>
-                    <FormattedMessage id='common.toggle.text' /> :
-                  </div>
-                  <div className={styles['card-value']}>
-                    <FormattedMessage id='connect.first.toggle.view.left' />
-                    <span className={styles['toggle-name']}>{ toggleName }</span>
-                    <FormattedMessage id='connect.first.toggle.view.right' />
-                    <span className={styles['toggle-key']}>{ toggleKey }</span>
-                  </div>
-                </div>
-              </div>
-            </>
-          )
-        }
+              </>
+            )
+          }
+        </div>
+        <div className={styles.steps}>
+          {
+            isStepLoading ? <Loading /> : (
+              <>
+                <StepFirst 
+                  currentStep={currentStep}
+                  currentSDK={currentSDK}
+                  clientAvailability={clientAvailability}
+                  saveStep={saveFirstStep}
+                  saveCurrentSDK={saveCurrentSDK}
+                  goBackToStep={goBackToStep}
+                />
+                <StepSecond 
+                  attributes={attributes}
+                  currentStep={currentStep}
+                  currentSDK={currentSDK}
+                  returnType={returnType}
+                  serverSdkKey={serverSdkKey}
+                  clientSdkKey={clientSdkKey}
+                  sdkVersion={sdkVersion}
+                  saveStep={saveSecondStep}
+                  goBackToStep={goBackToStep}
+                />
+                <StepThird 
+                  isLoading={isLoading}
+                  projectKey={projectKey}
+                  environmentKey={environmentKey}
+                  toggleKey={toggleKey}
+                  currentStep={currentStep}
+                  toggleAccess={toggleAccess}
+                  saveIsLoading={saveIsLoading}
+                  checkToggleStatus={checkToggleStatus}
+                />
+              </>
+            )
+          }
+        </div>
       </div>
-      <div className={styles.steps}>
-        {
-          isStepLoading ? <Loading /> : (
-            <>
-              <StepFirst 
-                currentStep={currentStep}
-                currentSDK={currentSDK}
-                clientAvailability={clientAvailability}
-                saveStep={saveFirstStep}
-                saveCurrentSDK={saveCurrentSDK}
-                goBackToStep={goBackToStep}
-              />
-              <StepSecond 
-                attributes={attributes}
-                currentStep={currentStep}
-                currentSDK={currentSDK}
-                returnType={returnType}
-                serverSdkKey={serverSdkKey}
-                clientSdkKey={clientSdkKey}
-                sdkVersion={sdkVersion}
-                saveStep={saveSecondStep}
-                goBackToStep={goBackToStep}
-              />
-              <StepThird 
-                isLoading={isLoading}
-                projectKey={projectKey}
-                environmentKey={environmentKey}
-                toggleKey={toggleKey}
-                currentStep={currentStep}
-                toggleAccess={toggleAccess}
-                saveIsLoading={saveIsLoading}
-                checkToggleStatus={checkToggleStatus}
-              />
-            </>
-          )
-        }
-      </div>
-    </div>
+    </ProjectLayout>
   );
 };
 
-export default Steps;
+export default ConnectSDK;
