--- ui/src/pages/getStarted/index.module.scss
@@ -1,16 +1,0 @@
-.get-started {
-  padding: 0;
-}
-
-.breadcrumb {
-  font-size: 12px;
-  line-height: 18px;
-  padding: 12px 24px 12px 48px;
-}
-
-.breadcrumb-icon {
-  font-size: 18px !important;
-  top: 2px;
-  position: relative;
-  color: #adb5bc;
-}
