--- ui/src/pages/getStarted/index.tsx
@@ -1,16 +1,0 @@
-import ProjectLayout from 'layout/projectLayout';
-import Steps from './components/Steps';
-import styles from './index.module.scss';
-
-const GetStarted = () => {
-
-  return (
-    <ProjectLayout>
-      <div className={styles['get-started']}>
-        <Steps />
-      </div>
-    </ProjectLayout>
-  );
-};
-
-export default GetStarted;
